requirements.txt
aio-pika==9.3.1
aiohttp==3.9.1
aiormq==6.7.7
aiosignal==1.3.1
alembic
amqp==5.2.0
attrs==23.1.0
bidict==0.22.1
billiard==4.2.0
bitarray==2.8.3
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-aliases==1.0.1
click-default-group==1.2.4
click-didyoumean==0.3.0
click-help-colors==0.9.4
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
hexbytes==0.3.1
idna==3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.4
lru-dict==1.2.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.3.0.post0
pamqp==3.2.1
parsimonious==0.9.0
pika==1.3.2
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pyunormalize==15.1.0
pywin32==306
redis==5.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
tabulate==0.8.10
toolz==0.12.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
web3==6.2.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.3
